chore(pg): address CodeRabbit review feedback
- Rewrite the changeset to be outcome-focused: lead with 'Fixed', describe
  the user-facing crash fix and auto-reconnect behavior, drop internal
  implementation details per .changeset path guidelines.
- Colocate the pool error listener tests with the source they cover
  (vector/, storage/, storage/db/) per repo testing guidelines, replacing
  the centralized src/pool-error-listener.test.ts. Assertions, setup, and
  cleanup are preserved verbatim.
